+++ js/data.js
@@ -1,0 +1,35 @@
+const OFFERS_LENGTH = 10;
+const OFFER_TITLE = [
+  'Большая уютная квартира',
+  'Маленькая неуютная квартира',
+  'Огромный прекрасный дворец',
+  'Маленький ужасный дворец',
+  'Красивый гостевой домик',
+  'Некрасивый негостеприимный домик',
+  'Уютное бунгало далеко от моря',
+  'Неуютное бунгало по колено в воде',
+];
+const OFFER_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const OFFER_TIMES = ['12:00', '13:00', '14:00'];
+const OFFER_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const OFFER_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {
+  OFFERS_LENGTH,
+  OFFER_TITLE,
+  OFFER_TYPES,
+  OFFER_TIMES,
+  OFFER_FEATURES,
+  OFFER_PHOTOS,
+};
